+++ src/Console/Interface/ExtendCommandsInterface.php
@@ -8,4 +8,9 @@ interface ExtendCommandsInterface
      * @return array
      */
     public function getCommandClass(): array;
+
+    /**
+     * @return string
+     */
+    public function getTypeCommands(): string;
 }