+++ src/Console/tests/CustomTest.php
@@ -8,14 +8,11 @@ use Solital\Core\Console\Interface\CommandInterface;
 class CustomTest extends Command implements CommandInterface
 {
     protected string $command = "create:testesecond";
-    protected array $arguments = [
-        "--argument-custom-second"
-    ];
+    protected array $arguments = ["name"];
     protected string $description = "Description second command";
 
     public function handle(object $arguments, object $options): mixed
     {
-        #$res = $this->getArguments();
-        var_dump($this->getCommand());
+        return $arguments->name;
     }
 }