+++ src/Console/tests/MyCommandsTest.php
@@ -8,13 +8,11 @@ use Solital\Core\Console\Interface\CommandInterface;
 class MyCommandsTest extends Command implements CommandInterface
 {
     protected string $command = "user:test";
-    protected array $arguments = [
-        "--argument-custom-second"
-    ];
+    protected array $arguments = ["argument_test"];
     protected string $description = "Description user command";
 
     public function handle(object $arguments, object $options): mixed
     {
-        return $this->getCommand();
+        return $this->getCommand() . " - Command OK";
     }
 }