+++ src/Kernel/Console/SolitalCommands.php
@@ -27,6 +27,11 @@ use Solital\Core\Kernel\Console\Commands\{
 class SolitalCommands implements ExtendCommandsInterface
 {
     /**
+     * @var string
+     */
+    protected string $type_commands = "Solital Commands";
+
+    /**
      * @var array
      */
     protected array $command_class = [
@@ -56,5 +61,13 @@ class SolitalCommands implements ExtendCommandsInterface
     public function getCommandClass(): array
     {
         return $this->command_class;
+    }
+
+    /**
+     * @return string
+     */
+    public function getTypeCommands(): string
+    {
+        return $this->type_commands;
     }
 }