+++ src/Kernel/Console/Templates/CommandTemplate.php
@@ -1,7 +1,9 @@
 <?php
+
 /**
  * @generated class generated using Vinci Console
  */
+
 namespace Solital\Console\Command;
 
 use Solital\Core\Console\{Command, MessageTrait};