+++ src/Kernel/Console/Templates/Commands/Config.php
@@ -12,10 +12,23 @@ class Config implements ExtendCommandsInterface
     protected array $command_class = [];
 
     /**
+     * @var string
+     */
+    protected string $type_commands = "";
+
+    /**
      * @return array
      */
     public function getCommandClass(): array
     {
         return $this->command_class;
+    }
+
+    /**
+     * @return string
+     */
+    public function getTypeCommands(): string
+    {
+        return $this->type_commands;
     }
 }