+++ src/Kernel/Console/Templates/ControllerTemplate.php
@@ -1,7 +1,9 @@
 <?php
+
 /**
  * @generated class generated using Vinci Console
  */
+
 namespace Solital\Components\Controller;
 
 use Solital\Core\Http\Controller\Controller;